# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenID Federation Resolver",
    "version": "1.0.0",
    "description": "Operations and federation APIs for this resolver instance."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "MetricsBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required for GET /metrics only when METRICS_TOKEN is set."
      },
      "OperatorBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_KEY or a PAT (oidf_pat_…) for /admin/v1, cache POST/DELETE, and configuration. Also accepted as X-API-Key. GET /api/v1 routes stay public. Minting PATs requires ENV API_KEY when it is set."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "X-API-Key: API_KEY or PAT for operator routes, or TA_API_KEY for trust-anchor register/unregister."
      },
      "TABearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "TA_API_KEY for POST /register-trust-anchor and DELETE registered-trust-anchors. TA_API_TOKEN is an alias. API_KEY is not accepted."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/resolve": {
      "get": {
        "tags": [
          "Federation"
        ],
        "summary": "Official federation resolve (OpenID Federation 1.0 \u00a78.3)",
        "parameters": [
          {
            "name": "sub",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trust_anchor",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "resolve-response+jwt"
          }
        }
      }
    },
    "/api/v1/federation_list": {
      "get": {
        "tags": [
          "Federation"
        ],
        "summary": "Federation member list",
        "parameters": [
          {
            "name": "trust_anchor",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member list"
          }
        }
      }
    },
    "/api/v1/collection": {
      "get": {
        "tags": [
          "Federation"
        ],
        "summary": "Entity collection (draft extension)",
        "parameters": [
          {
            "name": "trust_anchor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collected entities"
          }
        }
      }
    },
...
    "/api/v1/register-trust-anchor": {
      "post": {
        "tags": [
          "Trust anchors"
        ],
        "summary": "Register a trust anchor authorization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entity_id": {
                    "type": "string"
                  },
                  "entity_statement": {
                    "type": "string"
                  },
                  "registration_jwt": {
                    "type": "string"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered"
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        },
        "security": [
          {
            "TABearer": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
...
    "/api/v1/cache/clear-all": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "Clear entity and trust-chain caches",
        "responses": {
          "200": {
            "description": "Cleared"
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        },
        "security": [
          {
            "OperatorBearer": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/v1/cache/clear-entities": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "Clear entity cache",
        "responses": {
          "200": {
            "description": "Cleared"
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        },
        "security": [
          {
            "OperatorBearer": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/v1/cache/clear-chains": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "Clear trust-chain cache",
        "responses": {
          "200": {
            "description": "Cleared"
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        },
        "security": [
          {
            "OperatorBearer": []
          },
          {
            "ApiKey": []
          }
        ]
      }
    },
...
    "/api/v1/auth/verify": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Verify operator API_KEY without mutating state",
        "description": "Returns 200 when Authorization: Bearer or X-API-Key matches API_KEY. Used by OIDF Admin to validate take-control.",
        "security": [
          { "OperatorBearer": [] },
          { "ApiKey": [] }
        ],
        "responses": {
          "200": {
            "description": "Key accepted"
          },
          "401": {
            "description": "Missing or wrong API_KEY"
          }
        }
      }
    },
...
  }
}